+++ exec/parallel_with_approximation.R
@@ -9,13 +9,9 @@ parallel_approx_distance_matrix <- function(diagrams,other_diagrams = NULL,dim =
   # create cluster
   cl <- parallel::makeCluster(num_workers)
   doParallel::registerDoParallel(cl)
-  
-  # export TDApplied, and input parameters, to all cluster workers
-  parallel::clusterEvalQ(cl,c(library(TDApplied),library(foreach),library(iterators)))
-  parallel::clusterExport(cl,varlist = c("diagrams","dim","sigma","rho"),envir = environment())
-  
+
   # calculate distances in parallel
-  # clusters are closed
+  # clusters are closed if there is an error
   tryCatch(expr = {
     
     if(is.null(other_diagrams))
@@ -23,14 +19,39 @@ parallel_approx_distance_matrix <- function(diagrams,other_diagrams = NULL,dim =
       # not cross distance matrix, only need to compute the upper diagonal
       # since the matrix is symmetric
       d <- matrix(data = 0,nrow = length(diagrams),ncol = length(diagrams))
-      d_off_diag <- foreach::`%dopar%`(obj = foreach::foreach(r = iterators::iter(which(upper.tri(d),arr.ind = T),by = 'row'),.combine = c),ex = {TDApplied::diagram_distance(D1 = diagrams[[r[[1]]]],D2 = diagrams[[r[[2]]]],dim = dim,distance = "fisher",sigma = sigma,rho = rho)})
+      u <- which(upper.tri(d),arr.ind = T)
+      R <- lapply(X = 1:nrow(u),FUN = function(X){
+        
+        return(list(diagrams[[u[[X,1]]]],diagrams[[u[[X,2]]]]))
+        
+      })
+      
+      # remove diagrams to preserve memory
+      rm(diagrams)
+      
+      # calculate distances in parallel, export TDApplied to nodes
+      d_off_diag <- foreach::`%dopar%`(obj = foreach::foreach(r = R,.combine = c,.packages = c("TDApplied")),ex = {TDApplied::diagram_distance(D1 = r[[1]],D2 = r[[2]],dim = dim,distance = "fisher",sigma = sigma,rho = rho)})
+      
+      # store results in matrix
       d[upper.tri(d)] <- d_off_diag
       d[which(upper.tri(d),arr.ind = T)[,c("col","row")]] <- d_off_diag
       diag(d) <- rep(0,nrow(d))
     }else
     {
       # cross distance matrix, need to compute all entries
-      d <- foreach::`%dopar%`(foreach::`%:%`(foreach::foreach(r = 1:length(other_diagrams),.combine = cbind),foreach::foreach(X = 1:length(diagrams),.combine = c)),ex = {TDApplied::diagram_distance(D1 = other_diagrams[[r]],D2 = diagrams[[X]],dim = dim,distance = "fisher",sigma = sigma,rho = rho)})
+      u <- expand.grid(1:length(other_diagrams),1:length(diagrams))
+      R <- lapply(X = 1:nrow(u),FUN = function(X){
+        
+        return(list(other_diagrams[[u[X,1]]],diagrams[[u[X,2]]]))
+        
+      })
+      
+      # remove diagrams and other_diagrams to preserve memory
+      rm(list = c("diagrams","other_diagrams"))
+      
+      # store distance calculations in matrix
+      d[as.matrix(u)] <- foreach::`%dopar%`(foreach::foreach(r = R,.combine = cbind,.packages = c("TDApplied")),ex = {TDApplied::diagram_distance(D1 = r[[1]],D2 = r[[2]],dim = dim,distance = "fisher",sigma = sigma,rho = rho)})
+      
     }
     
   }, warning = function(w){warning(w)},