+++ NEWS.md
@@ -10,6 +10,7 @@ Please see those for more details.
 - added vignettes for speedups and HCP analysis (and removed those parts from the main vignette)
 - fixed issues with cv model fitting in diagram_ksvm
 - added automatic calculation of t parameters in diagram_ksvm
+- decreased memory load on parallel functions (except for permutation test loss function)
 
 # 2.0.4
 - fixed build issues related to use of suggested packages in tests, examples and vignettes